Dragging of fixtures

+++ src/ui/bracket_view.rs
@@ -1,13 +1,14 @@
-use log::{error};
+use std::collections::HashMap;
+
+use log::{error, debug};
 use wasm_bindgen::{JsCast};
-use web_sys::{HtmlElement, HtmlDivElement, MouseEvent, HtmlButtonElement};
+use web_sys::{HtmlElement, HtmlDivElement, MouseEvent, HtmlButtonElement, DomRect, window};
 
 use crate::{dom::{create_element}, model::tournament::{StageId, TournamentId, StageKind, FixtureId}, model::Model, ui::{UiElement, UiElementId, create_callback, EventList, Event}};
 
 use super::create_callback_with_arg;
 
 //TODO: drag fixture around
-//TODO: delete fixture
 
 pub struct BracketView {
     id: UiElementId,
@@ -16,9 +17,17 @@ pub struct BracketView {
     linked_outline_id: UiElementId,
 
     dom_root: HtmlElement,
+    fixture_divs: HashMap<FixtureId, HtmlDivElement>,
 
     #[allow(dyn_drop)]
     closures: Vec<Box<dyn Drop>>,
+
+    current_drag: Option<DragInfo>,
+}
+
+struct DragInfo {
+    fixture_id: FixtureId,
+    start_offset: (f64, f64),
 }
 
 impl BracketView {
@@ -55,18 +64,37 @@ impl BracketView {
         dom_root.set_inner_text("Bracket view!");
         dom_root.style().set_property("position", "relative").expect("Failed to set property"); // For children to be absolutely positioned relative to this.
 
+        #[allow(dyn_drop)]
+        let mut closures: Vec<Box<dyn Drop>> = vec![];
+
         let dblclick_closure = Box::new(create_callback_with_arg(move |model, ui, e| {
             if let Some(UiElement::BracketView(this)) = ui.get_element(id) {
                 this.on_background_dblclick(model, e);
             }
         }));
         dom_root.set_ondblclick(Some(dblclick_closure.as_ref().as_ref().unchecked_ref()));
-
-        #[allow(dyn_drop)]
-        let mut closures: Vec<Box<dyn Drop>> = vec![];
         closures.push(dblclick_closure); // Needs to be kept alive
 
-        let mut result = BracketView { id, tournament_id: None, stage_id: None, linked_outline_id, dom_root, closures };
+        let mousemove_closure = Box::new(create_callback_with_arg(move |model, ui, e| {
+            if let Some(UiElement::BracketView(this)) = ui.get_element(id) {
+                this.on_background_mousemove(model, e);
+            }
+        }));
+        dom_root.set_onmousemove(Some(mousemove_closure.as_ref().as_ref().unchecked_ref()));
+        closures.push(mousemove_closure); // Needs to be kept alive
+
+        let mouseup_closure = Box::new(create_callback_with_arg(move |model, ui, e| {
+            if let Some(UiElement::BracketView(this)) = ui.get_element_mut(id) {
+                this.on_background_mouseup(model, e);
+            }
+        }));
+        dom_root.set_onmouseup(Some(mouseup_closure.as_ref().as_ref().unchecked_ref()));
+        closures.push(mouseup_closure); // Needs to be kept alive
+
+
+
+        let mut result = BracketView { id, tournament_id: None, stage_id: None, linked_outline_id, dom_root, fixture_divs: HashMap::<FixtureId, HtmlDivElement>::new(), closures,
+            current_drag: None };
 
         result.refresh(model);
 
@@ -87,6 +115,8 @@ impl BracketView {
 
         self.dom_root.set_inner_html("");
         //TODO: delete closures?
+
+        self.fixture_divs.clear();
 
         if let (Some(tournament_id), Some(stage_id)) = (self.tournament_id, self.stage_id) {
             if let Some(stage) = model.get_stage(tournament_id, stage_id) {
@@ -111,8 +141,20 @@ impl BracketView {
                         delete_button.set_onclick(Some(click_closure.as_ref().as_ref().unchecked_ref()));
                         self.closures.push(click_closure); // Needs to be kept alive
 
+                        let drag_handle: HtmlElement = create_element("span");
+                        drag_handle.set_inner_text("::");
+                        new_div.append_child(&drag_handle).expect("Failed to append child");
+                        let mousedown_closure = Box::new(create_callback_with_arg(move |model, ui, e| {
+                            if let Some(UiElement::BracketView(this)) = ui.get_element_mut(id) {
+                                this.on_fixture_drag_handle_mousedown(model, fid, e);
+                            }
+                        }));
+                        drag_handle.set_onmousedown(Some(mousedown_closure.as_ref().as_ref().unchecked_ref()));
+                        self.closures.push(mousedown_closure); // Needs to be kept alive
 
                         self.dom_root.append_child(&new_div).expect("Failed to append child");
+
+                        self.fixture_divs.insert(fid, new_div);
                     }
                 }
             }
@@ -129,8 +171,50 @@ impl BracketView {
 
     fn on_delete_fixture_button_click(&self, model: &mut Model, fixture_id: FixtureId) {
         if let (Some(tournament_id), Some(stage_id)) = (self.tournament_id, self.stage_id) {
-            if let Err(_) = model.delete_fixture(tournament_id, stage_id, fixture_id) {
-                error!("Failed to delete fixture");
+            if window().unwrap().confirm_with_message(&format!("Are you sure you want to delete this fixture?")) == Ok(true) {
+                //TODO: also delete the associated match if any??
+                if let Err(_) = model.delete_fixture(tournament_id, stage_id, fixture_id) {
+                    error!("Failed to delete fixture");
+                }
+            }
+        }
+    }
+
+    fn on_fixture_drag_handle_mousedown(&mut self, model: &mut Model, fixture_id: FixtureId, e: MouseEvent) {
+        debug!("Dragging {fixture_id}");
+        if let Some(fixture_div) = self.fixture_divs.get(&fixture_id) {
+            let fixture_div_rect = fixture_div.get_bounding_client_rect();
+            self.current_drag = Some(DragInfo { fixture_id, start_offset: (e.client_x() as f64 - fixture_div_rect.left(), e.client_y() as f64 - fixture_div_rect.top()) });
+        }
+    }
+
+    fn on_background_mousemove(&self, model: &mut Model, e: MouseEvent) {
+        debug!("move to {} {}", e.offset_x(), e.offset_y());
+        if let Some(drag_info) = &self.current_drag {
+            if let Some(fixture_div) = self.fixture_divs.get(&drag_info.fixture_id) {
+                // Get the cursor position relative to the bracket view
+                // This won't be the same as e.offsetX/Y if we are currently dragging over a child element of the bracket view,
+                // as we receive the mousemove event via bubbling, so e.target will be the child element, and e.offsetX/Y will be relative to that instead.
+              //  let e_target = e.target().and_then(|t| t.dyn_into::<HtmlElement>().ok());
+              //  let target_rect =  e_target.as_ref().map_or(DomRect::new().unwrap(), |t| t.get_bounding_client_rect());
+                let root_rect = self.dom_root.get_bounding_client_rect();
+                // let x = e.offset_x() as f64 + target_rect.left() - root_rect.left();
+                // let y = e.offset_y() as f64 + target_rect.top() - root_rect.top();
+                let x = e.client_x() as f64 - root_rect.left() - drag_info.start_offset.0;
+                let y = e.client_y() as f64 - root_rect.top() - drag_info.start_offset.1;
+                // debug!("target rect = {},{}. root rect = {},{}. updated to {x} {y}", target_rect.left(), target_rect.top(), root_rect.left(), root_rect.top());
+
+                fixture_div.style().set_property("left", &x.to_string()).expect("Failed to set property");
+                fixture_div.style().set_property("top", &y.to_string()).expect("Failed to set property");
+            }
+        }
+    }
+
+    fn on_background_mouseup(&mut self, model: &mut Model, e: MouseEvent) {
+        if let Some(drag_info) = &self.current_drag {
+            if let Some(fixture_div) = self.fixture_divs.get(&drag_info.fixture_id) {
+                self.current_drag = None;
+                //TODO: save the new position in the model
             }
         }
     }
